rename api dirs
