+++ packages/elasticsearch-asset-apis/src/elasticsearch-date-slicer/helpers.ts
