+++ packages/elasticsearch-asset-apis/src/elasticsearch-id-slicer/helpers.ts
