+++ packages/elasticsearch-asset-apis/src/elasticsearch-id-slicer/index.ts
