+++ packages/elasticsearch-asset-apis/src/index.ts
@@ -3,5 +3,5 @@ export * from './window-state';
 export * from './spaces-api';
 export * from './elasticsearch-api';
 export * from './base-api';
-export * from './elasticsearch_date_slicer/helpers';
-export * from './elasticsearch_id_slicer/helpers';
+export * from './elasticsearch-date-slicer/helpers';
+export * from './elasticsearch-id-slicer/helpers';
